changed useThemeColor to getThemeColor
